udf: add a xwasm_b64 UDF language
Currently, when creating a WASM UDF, we have to first convert
the .wasm binary file to a text format, and the paste the entire
text file to the CQL query. This patch adds an option to pass
the binary file in a base64 format instead, decreasing the size
of CQL statements creating WASM UDFs, and removing the need
from UDF developers to download and use the 'wasm2wat' script
on .wasm files - the base64 representation can be generated using
a `base64 -w0` command.