compaction: create task manager's task for compaction group compaction
As compaction_manager::perform_major_compaction can be called from
many different places, we want to be able to run it with or without
task manager context, while using the same mechanism for both. Thus,
perform_major_compaction deals with bare task_impl_ptr, which is then
turned to task manager's task or not.