sstables: Add shard_owner_hint metadata
Shard owner hint is intended to help SSTable loader to quickly determine
which shard should open a particular SSTable.

Sharding metadata is more precise, but it can grow large (one of the
reasons we have to discard it from memory once we complete loading),
and the function to calculate the shard owning the SSTable from it
is not cheap. So we cannot use it to provide a hint.

Today, the hint is provided by looking at the SSTable generation
number.
But we want to decouple the hint from generation as tampering with
UUID may result in loss of uniqueness. CPUs today can grow to
even thousands of threads, and that would result in loss of up to
10 bits of information in the UUID.

Offline scripts can use this new metadata too, to quickly determine
which shard generated a particular SSTable.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>