statement_restrictions: add get_not_null_columns()
The IS NOT NULL restrictions are handled in a special way.
Instead of putting them together with other restrictions,
statement_restrictions collects all columns restricted
by IS NOT NULL and puts them in the _not_null_columns field.

Add a getter to access this set of columns.
The field is private, so it can't be accessed without
a function that explictily exposes it.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>