partition_version: handle multi-schema snapshots in merge_partition_versions
Each partition_version is allowed to have a different schema now.
As of this patch, all versions reachable from a snapshot/entry always
have the same schema, but this will change in an upcoming patch.
This commit prepares merge_partition_versions() for that.

When merging neighbouring versions, we usually the newer version into
the older one, mainly because of the educated guess that the newer
version is usually smaller, as it represents some incremental
update to the entry. Merging a smaller version into a bigger one is cheaper,
because less tree searches and updates have to be performed.

However, in the case of upgrades, we want all versions to be eventually
merged into a single version under the most recent schema, so that reads
don't have to perform row upgrades on the fly. That's why in case of upgrades
we have to merge the older version into the newer one.

After this patch, MVCC should be prepared for replacing the atomic schema
upgrade() of cache/memtable entries with a gentle upgrade().